requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
appdirs=1.4.4=pypi_0
asgiref=3.2.10=py_0
beckett=0.8.0=pypi_0
ca-certificates=2020.7.22=0
certifi=2020.6.20=py38_0
chardet=3.0.4=pypi_0
django=3.1.2=py_0
django-filter=2.4.0=pypi_0
djangorestframework=3.12.1=pypi_0
idna=2.8=pypi_0
inflect=0.2.5=pypi_0
krb5=1.17.1=h173b8e3_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpq=12.2=h20c2e04_0
libstdcxx-ng=9.1.0=hdf63c60_0
ncurses=6.2=he6710b0_1
openssl=1.1.1h=h7b6447c_0
pip=20.2.3=py38_0
pokebase=1.3.0=pypi_0
pokepy=0.6.1=pypi_0
psycopg2=2.8.5=py38h1ba5d50_0
python=3.8.5=h7579374_1
pytz=2020.1=py_0
readline=8.0=h7b6447c_0
requests=2.21.0=pypi_0
setuptools=50.3.0=py38hb0f4dca_1
six=1.10.0=pypi_0
sqlite=3.33.0=h62c20be_0
sqlparse=0.4.1=py_0
tk=8.6.10=hbc83047_0
urllib3=1.24.3=pypi_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
